requirements.txt
numpy==1.11.1
torch==0.2.0.post4
Pillow==5.0.0
torchvision==0.2.0
